Make routes relative and share common layout
