+++ src/components/Post/PostNav.jsx
@@ -6,6 +6,7 @@ import {
 } from "@mui/icons-material";
 import { IconButton, Typography } from "@mui/material";
 import React, { useState } from "react";
+import { ShareModal } from "../../reusableComponents";
 
 import BookmarkBorderIcon from "@mui/icons-material/BookmarkBorder";
 import BookmarksIcon from "@mui/icons-material/Bookmarks";
@@ -22,16 +23,14 @@ const PostNav = ({
   user,
   tempLikeCount,
   setisCommentOpen,
-  setLink,
-  setPostText,
-  shareModal,
 }) => {
   const { enqueueSnackbar } = useSnackbar();
   const [Open, setOpen] = useState(false);
+  const [openShareModal, setOpenShareModal] = useState(false);
   const [favoritePosts, setFavoritePosts] = useState(
     JSON.parse(localStorage.getItem("posts")) || [],
   );
-  const [isSaved, setisSaved] = useState(false);
+
   const navigate = useNavigate();
   const { isAnonymous } = user;
 
@@ -57,96 +56,93 @@ const PostNav = ({
     setFavoritePosts(JSON.parse(localStorage.getItem("posts")));
   };
 
-  const handleToggleFavorite = () => {
-    setisSaved(!isSaved);
-  };
 
-  const renderFavoriteIcon = () => {
-    if (isSaved) {
-      return <BookmarksIcon onClick={handleToggleFavorite} />;
-    } else {
-      return <BookmarkBorderIcon onClick={handleToggleFavorite} />;
-    }
-  };
 
   return (
-    <Flexbetween gap={!fullScreen && "1.6rem"} sx={{ marginInline: "auto" }}>
-      <Flexbetween
-        sx={{ cursor: "pointer" }}
-        onClick={() =>
-          isAnonymous ? navigate("/dummygram/signup") : likesHandler()
-        }
-      >
-        <IconButton>
-          {tempLikeCount.indexOf(user?.uid) != -1 ? (
-            <FavoriteOutlined
-              sx={{
-                color: "red",
-              }}
-            />
-          ) : (
-            <FavoriteBorderOutlined
+    <>
+      <Flexbetween gap={!fullScreen && "1.6rem"} sx={{ marginInline: "auto" }}>
+        <Flexbetween
+          sx={{ cursor: "pointer" }}
+          onClick={() =>
+            isAnonymous ? navigate("/dummygram/signup") : likesHandler()
+          }
+        >
+          <IconButton>
+            {tempLikeCount.indexOf(user?.uid) != -1 ? (
+              <FavoriteOutlined
+                sx={{
+                  color: "red",
+                }}
+              />
+            ) : (
+              <FavoriteBorderOutlined
+                style={{ color: "var(--post-nav-icons)" }}
+              />
+            )}
+          </IconButton>
+          <Typography fontSize={14} className="post-nav-item">
+            Like
+          </Typography>
+        </Flexbetween>
+
+        <Flexbetween
+          sx={{ cursor: "pointer" }}
+          onClick={() => {
+            isAnonymous ? navigate("/dummygram/signup") : setisCommentOpen(!Open);
+          }}
+        >
+          <IconButton sx={{ padding: "2px" }}>
+            <ChatBubbleOutlineRounded
               style={{ color: "var(--post-nav-icons)" }}
             />
-          )}
-        </IconButton>
-        <Typography fontSize={14} className="post-nav-item">
-          Like
-        </Typography>
-      </Flexbetween>
+          </IconButton>
+          <Typography fontSize={14} className="post-nav-item">
+            Comment
+          </Typography>
+        </Flexbetween>
 
-      <Flexbetween
-        sx={{ cursor: "pointer" }}
-        onClick={() => {
-          isAnonymous ? navigate("/dummygram/signup") : setisCommentOpen(!Open);
-        }}
-      >
-        <IconButton sx={{ padding: "2px" }}>
-          <ChatBubbleOutlineRounded
-            style={{ color: "var(--post-nav-icons)" }}
-          />
-        </IconButton>
-        <Typography fontSize={14} className="post-nav-item">
-          Comment
-        </Typography>
-      </Flexbetween>
+        <Flexbetween
+          sx={{ cursor: "pointer" }}
+          onClick={() => {
+            if (isAnonymous) {
+              navigate("/dummygram/signup");
+            } else {
+              setOpenShareModal(prev => !prev)
+            }
+          }}
+        >
+          <IconButton>
+            <ShareOutlined style={{ color: "var(--post-nav-icons)" }} />
+          </IconButton>
+          <Typography fontSize={14} className="post-nav-item">
+            Share
+          </Typography>
+        </Flexbetween>
 
-      <Flexbetween
-        sx={{ cursor: "pointer" }}
-        onClick={() => {
-          if (isAnonymous) {
-            navigate("/dummygram/signup");
-          } else {
-            setLink(`https://narayan954.github.io/dummygram/posts/${postId}`);
-            setPostText(caption);
-            shareModal(true);
-          }
-        }}
-      >
-        <IconButton>
-          <ShareOutlined style={{ color: "var(--post-nav-icons)" }} />
-        </IconButton>
-        <Typography fontSize={14} className="post-nav-item">
-          Share
-        </Typography>
+        <Flexbetween
+          sx={{ cursor: "pointer" }}
+          onClick={() => (isAnonymous ? navigate("/dummygram/signup") : save())}
+        >
+          <IconButton>
+            {favoritePosts.indexOf(postId) !== -1 ? (
+              <BookmarksIcon sx={{ color: "green" }} />
+            ) : (
+              <BookmarkBorderIcon style={{ color: "var(--post-nav-icons)" }} />
+            )}
+          </IconButton>
+          <Typography fontSize={14} className="post-nav-item">
+            Save
+          </Typography>
+        </Flexbetween>
       </Flexbetween>
-
-      <Flexbetween
-        sx={{ cursor: "pointer" }}
-        onClick={() => (isAnonymous ? navigate("/dummygram/signup") : save())}
-      >
-        <IconButton>
-          {favoritePosts.indexOf(postId) !== -1 ? (
-            <BookmarksIcon sx={{ color: "green" }} />
-          ) : (
-            <BookmarkBorderIcon style={{ color: "var(--post-nav-icons)" }} />
-          )}
-        </IconButton>
-        <Typography fontSize={14} className="post-nav-item">
-          Save
-        </Typography>
-      </Flexbetween>
-    </Flexbetween>
+      {openShareModal && (
+      <ShareModal
+        openShareModal={openShareModal}
+        setOpenShareModal={setOpenShareModal}
+        currentPostLink={`https://narayan954.github.io/dummygram/posts/${postId}`}
+        postText={caption}
+      />)}
+    </>
   );
 };
