+++ src/components/Suggestions/index.css
@@ -91,11 +91,6 @@
 }
 
 @media screen and (min-width: 800px) {
-  .suggestion-main-container {
-    margin-left: 8vw;
-    width: calc(100vw - 8vw);
-  }
-
   .suggestion-user-list-item {
     width: 300px;
     height: 250px;
@@ -103,13 +98,6 @@
 
   .suggestion-user-img {
     width: 120px;
-  }
-}
-
-@media screen and (min-width: 1200px) {
-  .suggestion-main-container {
-    margin-left: 17vw;
-    width: calc(100vw - 17vw);
   }
 }
