+++ src/index.css
@@ -84,6 +84,26 @@ body {
   overflow-x: auto;
 }
 
+.navbar_wrapper{
+  margin-top: 72px;
+}
+
+.sidebar_wrapper{
+  margin-left: 17vw;
+}
+
+@media screen and (max-width: 1200px) {
+  .sidebar_wrapper {
+    margin-left: 8vw;
+  }
+}
+
+@media screen and (max-width: 800px) {
+  .sidebar_wrapper {
+    margin-left: 0;
+  }
+}
+
 input:hover {
   cursor: text;
 }
@@ -102,8 +122,8 @@ input:focus {
 }
 
 /* Post style */
-.app__posts {
-  padding: 5.5em 17vw 20px 17vw;
+.app_posts {
+  padding: 1.5em 17vw 20px 0vw;
 }
 
 .app_posts_column {
@@ -111,7 +131,6 @@ input:focus {
   display: flex;
   flex-direction: column;
   justify-content: center;
-  padding: 5.5em 17vw 20px 17vw;
 }
 
 .rowConvert,
@@ -132,20 +151,18 @@ input:focus {
 }
 
 @media only screen and (max-width: 1300px) {
-  .app__posts,
-  .home-posts-container .app_posts_column,
   .home-posts-container .app_posts {
-    padding: 5.5em 15px 20px 9vw;
+    padding: 1.5em 15px 20px 10px;
   }
   .post-page-container {
-    padding: 5.5em 15px 15px 8vw;
+    padding: 1.5em 15px 15px 8vw;
   }
   .post-card {
     margin: 50px auto !important;
   }
 }
 @media only screen and (max-width: 800px) {
-  .app__posts,
+  .app_posts,
   .home-posts-container .app_posts_column {
     padding: 5.5em 15px 5.5em 15px;
     display: flex;
@@ -155,7 +172,7 @@ input:focus {
   }
   .post-page-container {
     padding: 0;
-    padding-top: 5.5em;
+    padding-top: 1em;
   }
 }
 
@@ -165,7 +182,7 @@ input:focus {
   }
 }
 @media only screen and (max-width: 700px) {
-  .app__posts {
+  .app_posts {
     padding-inline: auto;
   }
 }
@@ -269,8 +286,8 @@ input:focus {
   .app__newpost__button {
     left: -6%;
   }
-  .home-posts-container .app_posts_column {
-    padding: 5.5em 15px 5em 15px;
+  .home-posts-container .app_posts {
+    padding: 1.5em 15px 5em 15px;
   }
   .search-modal {
     width: 80vw;
@@ -309,14 +326,14 @@ input:focus {
 
 /* Extra large devices (large laptops and desktops, 1200px and up) */
 @media only screen and (min-width: 800px) {
-  .app__posts {
+  .app_posts {
     display: flex;
     flex-wrap: wrap;
     gap: 2rem;
     justify-content: center;
   }
   .post-page-container {
-    padding: 5.5em 15px 15px 17vw;
+    padding: 0 15px 15px 0;
   }
   .post-card {
     margin: 50px auto !important;
@@ -327,9 +344,6 @@ input:focus {
 
   .app__newpost__button {
     left: -1.2%;
-  }
-  .home-posts-container .app_posts_column {
-    padding: 5.5em 17vw 20px 17vw;
   }
 }
